requirements.txt
aiofiles==0.7.0
boxsdk==2.13.0
dnspython==2.1.0
fastapi==0.65.3
gunicorn==20.1.0
pymongo==3.12.0
python-dotenv==0.19.0
uvicorn[standard]==0.15.0
